pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cusflow"
version = "1.0.0"
description = "Production-ready Learning-to-Rank recommendation system with GenAI features"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "CusFlow Team" }
]
keywords = [
    "machine-learning",
    "recommendations",
    "learning-to-rank",
    "lambdamart",
    "genai",
    "embeddings",
    "fastapi",
    "redis"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "redis>=5.0.0",
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "scikit-learn>=1.4.0",
    "lightgbm>=4.3.0",
    "xgboost>=2.0.0",
    "sentence-transformers>=2.3.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "faiss-cpu>=1.7.4",
    "httpx>=0.26.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "joblib>=1.3.0",
    "tqdm>=4.66.0",
    "polars>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
]
gpu = [
    "faiss-gpu>=1.7.4",
]

[project.scripts]
cusflow = "src.cli:app"

[project.urls]
Homepage = "https://github.com/yourusername/cusflow"
Documentation = "https://github.com/yourusername/cusflow#readme"
Repository = "https://github.com/yourusername/cusflow"
Issues = "https://github.com/yourusername/cusflow/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501", "B008", "C901"]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"
